+++ src/Player.cc
@@ -10,8 +10,8 @@
 
 Player::Player(const SDL_Rect& hitbox)
     :  GameObject(AssetManager::getInstance().getTexture("player"), hitbox, 1.0),
-       pos_x{ static_cast<double>(hitbox.x) },
-       pos_y{ static_cast<double>(hitbox.y) } {}
+       pos_x{ static_cast<float>(hitbox.x) },
+       pos_y{ static_cast<float>(hitbox.y) } {}
 
 
 bool Player::hasSpecialWeapon() const
